move all type interfaces into /lib/types
