+++ salon-frontend/src/lib/hooks/useUsers.ts
@@ -1,7 +1,6 @@
 import { collection, doc, getDoc, getDocs, query, where } from "firebase/firestore";
 import { firebaseDb } from "@/lib/firebase";
-import { SalonUser } from "@/lib/hooks/useUserProfile";
-import { SalonRole } from "@/lib/types/types";
+import { SalonRole, SalonUser } from "@/lib/types/types";
 
 interface GetNameFromUserIdProps {
     userId: string;
@@ -29,12 +28,6 @@ interface FetchAllUsersReturn {
 interface FetchFromPhoneOrEmailProps {
     phoneNumber: string;
     email: string;
-}
-export interface SalonEmployee {
-    id: string;
-    firstName: string;
-    lastName: string;
-    role: SalonRole;
 }
 
 interface UseUserReturn {
